+++ src/main/java/org/dx42/needy/report/AbstractReport.groovy
@@ -28,6 +28,7 @@ abstract class AbstractReport implements Report {
 	String outputFile
 	String includeApplications 
 	String excludeApplications
+	String notesHtml
 	
 	protected Closure getDate = { new Date() }