+++ src/test/java/org/dx42/needy/report/ByArtifactHtmlReportTest.groovy
@@ -40,15 +40,23 @@ class ByArtifactHtmlReportTest extends AbstractTestCase {
 
 	private static final String EXPECTED_REPORT_TEXT = """
 		<!DOCTYPE html><html><head><title>Needy Dependency Report: Dependency Report</title><meta http-equiv="Content-Type" content="text/html"><style type='text/css'>$CSS</style></head>
-		<body><h1>Needy Dependency Report</h1><div class='metadata'><table><tr><td class='em'>Report Title:</td><td class='reportTitle'>Dependency Report</td></tr><tr><td class='em'>Timestamp:</td><td>$TIMESTAMP_STRING</td></tr><tr><td class='em'>Generated With:</td><td><a href='https://github.com/dx42/Needy'>Needy v$VERSION</a></td></tr></table></div><div class='summary'><h2>Dependencies</h2><table><tr class='tableHeader'><th>#</th><th>Group</th><th>Name</th><th>Version</th><th>Applications</th></tr><tr><td>1</td><td>log4j</td><td>log4j</td><td>1.2.14</td><td class='applicationNames'>Sample1, Sample_Two, Third</td></tr><tr><td>2</td><td>log4j-extra</td><td>stuff</td><td>1.0</td><td class='applicationNames'>Third</td></tr><tr><td>3</td><td>org.hibernate</td><td>hibernate-core</td><td>3.1</td><td class='applicationNames'>Sample1</td></tr><tr><td>4</td><td>org.other</td><td>service</td><td>2.0</td><td class='applicationNames'>Third</td></tr></table></div></body></html>
+		<body><h1>Needy Dependency Report</h1><div class='metadata'><table><tr><td class='em'>Report Title:</td><td class='reportTitle'>Dependency Report</td></tr>
+		<tr><td class='em'>Timestamp:</td><td>$TIMESTAMP_STRING</td></tr>
+		<tr><td class='em'>Generated With:</td><td><a href='https://github.com/dx42/Needy'>Needy v$VERSION</a></td></tr></table></div>
+		<div class='summary'><h2>Dependencies</h2><table><tr class='tableHeader'><th>#</th><th>Group</th><th>Name</th><th>Version</th><th>Applications</th></tr>
+		<tr><td>1</td><td>log4j</td><td>log4j</td><td>1.2.14</td><td class='applicationNames'>Sample1, Sample_Two, Third</td></tr>
+		<tr><td>2</td><td>log4j-extra</td><td>stuff</td><td>1.0</td><td class='applicationNames'>Third</td></tr>
+		<tr><td>3</td><td>org.hibernate</td><td>hibernate-core</td><td>3.1</td><td class='applicationNames'>Sample1</td></tr>
+		<tr><td>4</td><td>org.other</td><td>service</td><td>2.0</td><td class='applicationNames'>Third</td></tr>
+		</table></div></body></html>
 		"""
 
 	private static final String EXPECTED_REPORT_TEXT_NO_THIRD = """
 		<!DOCTYPE html><html><head><title>Needy Dependency Report: Dependency Report</title><meta http-equiv="Content-Type" content="text/html"><style type='text/css'>$CSS</style></head>
 		<body><h1>Needy Dependency Report</h1><div class='metadata'><table><tr><td class='em'>Report Title:</td><td class='reportTitle'>Dependency Report</td></tr>
 		<tr><td class='em'>Timestamp:</td><td>$TIMESTAMP_STRING</td></tr>
-		<tr><td class='em'>Generated With:</td><td><a href='https://github.com/dx42/Needy'>Needy v$VERSION</a></td></tr>
-		</table></div><div class='summary'><h2>Dependencies</h2><table><tr class='tableHeader'><th>#</th>
+		<tr><td class='em'>Generated With:</td><td><a href='https://github.com/dx42/Needy'>Needy v$VERSION</a></td></tr></table>
+		</div><div class='summary'><h2>Dependencies</h2><table><tr class='tableHeader'><th>#</th>
 		<th>Group</th><th>Name</th><th>Version</th><th>Applications</th></tr>
 		<tr><td>1</td><td>log4j</td><td>log4j</td><td>1.2.14</td><td class='applicationNames'>Sample1, Sample_Two</td></tr>
 		<tr><td>2</td><td>org.hibernate</td><td>hibernate-core</td><td>3.1</td><td class='applicationNames'>Sample1</td></tr>
@@ -100,6 +108,15 @@ class ByArtifactHtmlReportTest extends AbstractTestCase {
 		report.excludeApplications = "Th*"
 		report.writeReport(writer, DEPENDENCIES)
 		assertSameXml(writer.toString(), EXPECTED_REPORT_TEXT_NO_THIRD)
+	}
+	
+	@Test
+	void test_writeReport_notesHtml() {
+		String notesHtml = "<h2>Notes</h2><p>Some text</p>"
+		report.notesHtml = notesHtml
+		report.writeReport(writer, DEPENDENCIES)
+		String expected = EXPECTED_REPORT_TEXT.replace("<div class='summary'>", notesHtml + "<div class='summary'>")
+		assertSameXml(writer.toString(), expected)
 	}
 	
 	@Test